Unify handling of SVG and other images in RST.
ReadMeHTMLTranslator was overriden to not to render object tags for SVG.
Unfortunately it produced different results than for PNG or other
image types.

With this commit, HTML code for SVG images will be same as for other
images. It means that it won't have width, height or align attributes.
But it will be better to address this issue for all images types at once.