chore: remove unused styles parameter
Surfaced via `mypy`, recommended adding a type to the empty list.

The list was originally empty back in 0.1.0.

Instead of adding a type, remove the constant, and the code that uses it
from `clean()` - as it was partially reverted in #121.

The default `ALLOWED_STYLES` in the underlying library is `[]`.

Related: https://github.com/pypa/readme_renderer/pull/114#discussion_r218405013

Signed-off-by: Mike Fiedler <miketheman@gmail.com>