fix: allow math in rst
The first step to allowing math directives to be rendered on PyPI is
to enable the directive's warning to pass.

`docutils` tells us to pass a path to the `math_output` setting.
Since we `clean` the output, we can pass anything we want, as control
of the CSS will be handed on PyPI.

Refs: https://docutils.sourceforge.io/docs/user/config.html#math-output

This does **not** solve for Markdown math.

Resolves #115
Resolves #172

Signed-off-by: Mike Fiedler <miketheman@gmail.com>