Multiple pyproject.toml fixes (#251)
* Remove redundant wheel dep from pyproject.toml

Remove the redundant `wheel` dependency, as it is added by the backend
automatically.  Listing it explicitly in the documentation was
a historical mistake and has been fixed since, see:
https://github.com/pypa/setuptools/commit/f7d30a9529378cf69054b5176249e5457aaf640a

Signed-off-by: Michał Górny <mgorny@gentoo.org>

* Remove install-time dependencies from pyproject.toml requires

Remove the install-time dependencies from pyproject.toml `requires` key.
This key is used to specify the packages that are needed to build
the wheel, and FWICS none of these packages are actually needed
throughout the run of `setup.py`.  I can also confirm that a pure venv
`python -m build -w` works fine without them.

Signed-off-by: Michał Górny <mgorny@gentoo.org>

* Use the non-legacy setuptools backend in pyproject.toml

Use the regular `setuptools.build_meta` backend rather than
the `__legacy__` backend.  The latter was never meant to be used
explicitly in `pyproject.toml` files but rather as implicit fallback
in tools such as pip(1).  See e.g.:
https://github.com/pypa/setuptools/issues/1689

Signed-off-by: Michał Górny <mgorny@gentoo.org>

Co-authored-by: Dustin Ingram <di@users.noreply.github.com>